# src/docs/swagger.yaml
basePath: /api
definitions:
  models.Food:
    properties:
      createdAt:
        example: "2024-11-20T02:18:56.744307Z"
        type: string
      id:
        example: 4e51bdb9-c75a-4198-8f2d-9695e3ffaa83
        type: string
      name:
        example: Macarrão à Milanesa
        type: string
      price:
        example: 15.47
        type: number
      updatedAt:
        example: "2024-11-20T02:18:56.744307Z"
        type: string
    type: object
  request.Food:
    properties:
      name:
        example: Arroz com Passas
        type: string
      price:
        example: 14.75
        type: number
    required:
    - name
    - price
    type: object
  response.ApiError:
    properties:
      error:
        type: string
    type: object
  response.SearchResult:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Food'
        type: array
      page:
        example: 1
        type: integer
      perPage:
        example: 20
        type: integer
      total:
        example: 15
        type: integer
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact:
    email: lgcs10@aluno.ifal.edu.br
    name: Luis Gabriel da Costa Silva
  description: API simples para demonstrar o uso de containers Docker e suas integrações
    usando Docker Compose.
  title: API de Comida.
  version: "1.0"
paths:
  /foods:
    get:
      description: Retorna a lista de comidas cadastradas.
      parameters:
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Número de resultados por página
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SearchResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiError'
      summary: Lista Comidas.
      tags:
      - Comida
    post:
      consumes:
      - application/json
      description: Cadastra uma nova Comida.
      parameters:
      - description: Informações da nova Comida
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/request.Food'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Food'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiError'
      summary: Adiciona Comida.
      tags:
      - Comida
  /foods/{id}:
    delete:
      description: Deleta uma Comida existente.
      parameters:
      - description: ID da Comida
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiError'
      summary: Remove Comida.
      tags:
      - Comida
    put:
      consumes:
      - application/json
      description: Atualiza uma Comida existente.
      parameters:
      - description: ID da Comida
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Novas Informações da Comida
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/request.Food'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Food'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiError'
      summary: Atualiza Comida.
      tags:
      - Comida
swagger: "2.0"